pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raglab"
version = "0.1.0"
description = "A complete RAG pipeline for learning and experimentation with advanced retrieval techniques"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["rag", "retrieval-augmented-generation", "llm", "vector-database", "graphrag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "weaviate-client>=4.18.0",
    "neo4j>=6.0.0",
    "graphdatascience>=1.18",
    "docling>=2.60.0",
    "spacy>=3.8.0",
    "tiktoken>=0.12.0",
    "openai>=2.8.0",
    "httpx>=0.28.0",
    "instructor>=1.13.0",
    "pydantic>=2.12.0",
    "numpy>=1.24.0",
    "scipy>=1.15.0",
    "scikit-learn>=1.5.0",
    "umap-learn>=0.5.9",
    "ragas>=0.3.9",
    "datasets>=4.4.0",
    "streamlit>=1.52.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.urls]
Repository = "https://github.com/fliperbaker/raglab"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
